length-service/requirements.txt
Flask==2.3.2
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
prometheus-client
